refactor : 검증 테스트 패키지 변경
